[CHG] * removed quarkus quinoa * updated vehicle

+++ src/main/kotlin/org/lamiey/service/VehicleService.kt
@@ -3,28 +3,28 @@ package org.lamiey.service
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.ws.rs.WebApplicationException
 import jakarta.ws.rs.core.Response
+import org.lamiey.dto.VehicleDTO
 import org.lamiey.entity.Vehicle
 import org.lamiey.repository.VehicleRepository
-import org.lamiey.dto.VehicleDTO
 
 @ApplicationScoped
-class VehicleService(
-    private val vehicleRepository: VehicleRepository
-) {
-    //region GETTER
-    fun getVehicles(pageNumber: Int): List<Vehicle> = vehicleRepository.getAllVehicles(pageNumber)
+class VehicleService(private val vehicleRepository: VehicleRepository) {
+    // region GETTER
+    fun getVehicles(pageNumber: Int, pageSize: Int): List<Vehicle> =
+            vehicleRepository.getAllVehicles(pageNumber, pageSize)
 
     fun getVehicleById(id: Long): Vehicle? = vehicleRepository.findById(id)
-    //endregion
+    // endregion
 
-    //region CRUD
+    // region CRUD
     fun createVehicle(vehicleDTO: VehicleDTO): Vehicle {
-        val vehicle = Vehicle().apply {
-            licensePlate = vehicleDTO.licensePlate
-            name = vehicleDTO.name
-            manufacturer = vehicleDTO.manufacturer
-            model = vehicleDTO.model
-        }
+        val vehicle =
+                Vehicle().apply {
+                    licensePlate = vehicleDTO.licensePlate
+                    name = vehicleDTO.name
+                    manufacturer = vehicleDTO.manufacturer
+                    model = vehicleDTO.model
+                }
 
         vehicleRepository.persist(vehicle)
 
@@ -34,14 +34,17 @@ class VehicleService(
     fun updateVehicle(vehicleDTO: VehicleDTO): Vehicle {
         val vehicle: Vehicle? = vehicleDTO.id?.let { getVehicleById(it) }
 
-        vehicle ?: throw WebApplicationException(Response.status(Response.Status.NOT_FOUND).build())
+        if (vehicle == null)
+                throw WebApplicationException(Response.status(Response.Status.NOT_FOUND).build())
 
-        vehicleRepository.persist(vehicle.apply {
-            licensePlate = vehicleDTO.licensePlate
-            name = vehicleDTO.name
-            manufacturer = vehicleDTO.manufacturer
-            model = vehicleDTO.model
-        })
+        vehicleRepository.persist(
+                vehicle.apply {
+                    licensePlate = vehicleDTO.licensePlate
+                    name = vehicleDTO.name
+                    manufacturer = vehicleDTO.manufacturer
+                    model = vehicleDTO.model
+                }
+        )
 
         return vehicle
     }
@@ -49,9 +52,10 @@ class VehicleService(
     fun deleteVehicle(vehicleDTO: VehicleDTO) {
         val vehicle: Vehicle? = vehicleDTO.id?.let { vehicleRepository.findById(it) }
 
-        vehicle ?: throw WebApplicationException(Response.status(Response.Status.NOT_FOUND).build())
+        if (vehicle == null)
+                throw WebApplicationException(Response.status(Response.Status.NOT_FOUND).build())
 
         vehicleRepository.delete(vehicle)
     }
-    //endregion
+    // endregion
 }
